# docs/swagger/swagger.yaml
definitions:
    github_com_rickferrdev_gostock_internal_core_ports.Code:
        enum:
            - INTERNAL_SERVER_ERROR
            - BAD_REQUEST
            - RESOURCE_NOT_FOUND
            - UNAUTHENTICATED
            - PERMISSION_DENIED
            - RESOURCE_CONFLICT
            - USER_NOT_FOUND
            - INVALID_PASSWORD
            - EMAIL_ALREADY_REGISTERED
            - AUTH_TOKEN_EXPIRED
            - DATABASE_CONNECTION_FAILED
            - OPERATION_TIMEOUT
        type: string
        x-enum-varnames:
            - CodeInternal
            - CodeBadRequest
            - CodeNotFound
            - CodeUnauthenticated
            - CodeForbidden
            - CodeConflict
            - CodeUserNotFound
            - CodeInvalidPassword
            - CodeEmailRegistered
            - CodeTokenExpired
            - CodeDatabaseError
            - CodeTimeout
    github_com_rickferrdev_gostock_internal_core_ports.DTOProduct:
        properties:
            id:
                type: string
            name:
                type: string
            price:
                type: integer
            qtd:
                type: integer
            stock_id:
                type: string
        type: object
    github_com_rickferrdev_gostock_internal_core_ports.DTORequestCreateProduct:
        properties:
            name:
                minLength: 2
                type: string
            price:
                minimum: 1
                type: integer
            qtd:
                minimum: 0
                type: integer
            stock_id:
                type: string
        required:
            - name
            - price
            - stock_id
        type: object
    github_com_rickferrdev_gostock_internal_core_ports.DTORequestCreateStock:
        properties:
            capacity:
                minimum: 1
                type: integer
            name:
                minLength: 3
                type: string
        required:
            - capacity
            - name
        type: object
    github_com_rickferrdev_gostock_internal_core_ports.DTORequestUpdateProduct:
        properties:
            name:
                minLength: 2
                type: string
            price:
                minimum: 1
                type: integer
            qtd:
                minimum: 0
                type: integer
            stock_id:
                type: string
        type: object
    github_com_rickferrdev_gostock_internal_core_ports.DTORequestUpdateStock:
        properties:
            capacity:
                minimum: 1
                type: integer
            name:
                minLength: 3
                type: string
        required:
            - capacity
            - name
        type: object
    github_com_rickferrdev_gostock_internal_core_ports.DTOResponseAllStock:
        properties:
            data:
                items:
                    $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTOStock"
                type: array
        type: object
    github_com_rickferrdev_gostock_internal_core_ports.DTOResponseByIDProduct:
        properties:
            data:
                $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTOProduct"
        type: object
    github_com_rickferrdev_gostock_internal_core_ports.DTOResponseByIDStock:
        properties:
            data:
                $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTOStock"
        type: object
    github_com_rickferrdev_gostock_internal_core_ports.DTOResponseByStockIDProduct:
        properties:
            data:
                items:
                    $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTOProduct"
                type: array
        type: object
    github_com_rickferrdev_gostock_internal_core_ports.DTOResponseCreateProduct:
        properties:
            data:
                $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTOProduct"
        type: object
    github_com_rickferrdev_gostock_internal_core_ports.DTOResponseCreateStock:
        properties:
            data:
                $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTOStock"
        type: object
    github_com_rickferrdev_gostock_internal_core_ports.DTOStock:
        properties:
            capacity:
                type: integer
            id:
                type: string
            name:
                type: string
            used_capacity:
                type: integer
        type: object
    github_com_rickferrdev_gostock_internal_core_ports.Error:
        properties:
            code:
                $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Code"
            err: {}
            message:
                $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Message"
            status:
                type: integer
        type: object
    github_com_rickferrdev_gostock_internal_core_ports.Message:
        enum:
            - an unexpected internal error occurred
            - the requested resource was not found
            - the provided data is invalid
            - you must be authenticated to access this resource
            - you do not have permission to perform this action
            - user with these details already exists
            - invalid request parameters
            - cannot remove a stock that still has registered products
            - the resource's capacity was exceeded
            - error processing the standard submission message
        type: string
        x-enum-varnames:
            - MsgInternal
            - MsgNotFound
            - MsgInvalidInput
            - MsgUnauthorized
            - MsgForbidden
            - MsgUserConflict
            - MsgBadRequest
            - MsgStockNotEmpty
            - MsgCapacityExceeded
            - MsgProcessingMessage
info:
    contact: {}
paths:
    /products:
        post:
            consumes:
                - application/json
            description: Creates a new product and associates it with a stock
            parameters:
                - description: Product payload
                  in: body
                  name: body
                  required: true
                  schema:
                      $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTORequestCreateProduct"
            produces:
                - application/json
            responses:
                "201":
                    description: Created
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTOResponseCreateProduct"
                "400":
                    description: Invalid request body
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "500":
                    description: Internal server error
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
            summary: Create a product
            tags:
                - products
    /products/{id}:
        delete:
            consumes:
                - application/json
            description: Permanently removes a product by its ID
            parameters:
                - description: Product ID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: Product deleted successfully
                "400":
                    description: Missing or invalid product ID
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "404":
                    description: Product not found
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "500":
                    description: Internal server error
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
            summary: Delete a product
            tags:
                - products
        get:
            consumes:
                - application/json
            description: Returns a single product by its unique identifier
            parameters:
                - description: Product ID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTOResponseByIDProduct"
                "400":
                    description: Missing or invalid product ID
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "404":
                    description: Product not found
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "500":
                    description: Internal server error
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
            summary: Get a product by ID
            tags:
                - products
        put:
            consumes:
                - application/json
            description: Updates an existing product's fields by its ID
            parameters:
                - description: Product ID
                  in: path
                  name: id
                  required: true
                  type: string
                - description: Updated product payload
                  in: body
                  name: body
                  required: true
                  schema:
                      $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTORequestUpdateProduct"
            produces:
                - application/json
            responses:
                "200":
                    description: Product updated successfully
                "400":
                    description: Missing or invalid product ID / request body
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "404":
                    description: Product not found
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "500":
                    description: Internal server error
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
            summary: Update a product
            tags:
                - products
    /stocks:
        get:
            consumes:
                - application/json
            description:
                Returns a list of all available stocks with their current capacity
                usage
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTOResponseAllStock"
                "500":
                    description: Internal server error
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
            summary: List all stocks
            tags:
                - stocks
        post:
            consumes:
                - application/json
            description: Creates a new stock with the given name and total capacity
            parameters:
                - description: Stock payload
                  in: body
                  name: body
                  required: true
                  schema:
                      $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTORequestCreateStock"
            produces:
                - application/json
            responses:
                "201":
                    description: Created
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTOResponseCreateStock"
                "400":
                    description: Invalid request body
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "500":
                    description: Internal server error
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
            summary: Create a stock
            tags:
                - stocks
    /stocks/{id}:
        delete:
            consumes:
                - application/json
            description: Permanently removes a stock by its ID
            parameters:
                - description: Stock ID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: Stock deleted successfully
                "400":
                    description: Missing or invalid stock ID
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "404":
                    description: Stock not found
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "500":
                    description: Internal server error
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
            summary: Delete a stock
            tags:
                - stocks
        get:
            consumes:
                - application/json
            description:
                Returns a single stock with its current occupancy by its unique
                identifier
            parameters:
                - description: Stock ID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTOResponseByIDStock"
                "400":
                    description: Missing or invalid stock ID
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "404":
                    description: Stock not found
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "500":
                    description: Internal server error
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
            summary: Get a stock by ID
            tags:
                - stocks
        put:
            consumes:
                - application/json
            description: Updates the name and/or capacity of an existing stock by its ID
            parameters:
                - description: Stock ID
                  in: path
                  name: id
                  required: true
                  type: string
                - description: Updated stock payload
                  in: body
                  name: body
                  required: true
                  schema:
                      $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTORequestUpdateStock"
            produces:
                - application/json
            responses:
                "200":
                    description: Stock updated successfully
                "400":
                    description: Missing or invalid stock ID / request body
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "404":
                    description: Stock not found
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "500":
                    description: Internal server error
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
            summary: Update a stock
            tags:
                - stocks
    /stocks/{id}/products:
        get:
            consumes:
                - application/json
            description: Returns all products associated with a given stock ID
            parameters:
                - description: Stock ID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.DTOResponseByStockIDProduct"
                "400":
                    description: Missing or invalid stock ID
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "404":
                    description: Stock not found
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
                "500":
                    description: Internal server error
                    schema:
                        $ref: "#/definitions/github_com_rickferrdev_gostock_internal_core_ports.Error"
            summary: List products by stock
            tags:
                - products
swagger: "2.0"